server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.7
cloudinary==1.35.0
django-cloudinary-storage==0.3.0
celery==5.3.4
redis==5.0.1
django-storages==1.14.2
whitenoise==6.6.0
gunicorn==21.2.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0 
twilio==9.2.3